refactor: use u16 instead of usize to represent progress bar

+++ src/events/payload.rs
@@ -243,7 +243,7 @@ pub enum EventType {
         /// Progress.
         ///
         /// 0=error, 1-999=progress in permille, 1000=success and done
-        progress: usize,
+        progress: u16,
 
         /// Progress comment or error, something to display to the user.
         comment: Option<String>,
@@ -253,7 +253,7 @@ pub enum EventType {
     ///
     /// @param data1 (usize) 0=error, 1-999=progress in permille, 1000=success and done
     /// @param data2 0
-    ImexProgress(usize),
+    ImexProgress(u16),
 
     /// A file has been exported. A file has been written by imex().
     /// This event may be sent multiple times by a single call to imex().
@@ -280,7 +280,7 @@ pub enum EventType {
         chat_type: Chattype,
 
         /// Progress, always 1000.
-        progress: usize,
+        progress: u16,
     },
 
     /// Progress information of a secure-join handshake from the view of the joiner
@@ -295,7 +295,7 @@ pub enum EventType {
         /// 400=vg-/vc-request-with-auth sent, typically shown as "alice@addr verified, introducing myself."
         /// (Bob has verified alice and waits until Alice does the same for him)
         /// 1000=vg-member-added/vc-contact-confirm received
-        progress: usize,
+        progress: u16,
     },
 
     /// The connectivity to the server changed.
